requirements.txt
openvino>=2024.0
openvino-dev>=2024.0
numpy>=1.24.0
tensorflow>=2.12.0
tensorflow-hub>=0.16.0
Pillow>=9.0.0
setuptools
